Recompile 'email.x', add doudle id

+++ rpc-email/email.h
@@ -17,7 +17,7 @@ extern "C" {
 #define MAX_B 240
 
 struct temail {
-	long id;
+	double id;
 	char sender[MAX_SR];
 	char recipient[MAX_SR];
 	char subject[MAX_SR];
